Fix comparison for reference product types
