+++ src/integration-test/java/com/commercetools/sync/integration/externalsource/producttypes/ProductTypeSyncIT.java
@@ -40,14 +40,18 @@ import io.sphere.sdk.client.SphereClient;
 import io.sphere.sdk.commands.UpdateAction;
 import io.sphere.sdk.models.EnumValue;
 import io.sphere.sdk.models.LocalizedEnumValue;
+import io.sphere.sdk.models.LocalizedString;
 import io.sphere.sdk.models.TextInputHint;
 import io.sphere.sdk.products.attributes.AttributeConstraint;
+import io.sphere.sdk.products.attributes.AttributeDefinition;
+import io.sphere.sdk.products.attributes.AttributeDefinitionBuilder;
 import io.sphere.sdk.products.attributes.AttributeDefinitionDraft;
 import io.sphere.sdk.products.attributes.AttributeDefinitionDraftBuilder;
 import io.sphere.sdk.products.attributes.AttributeDefinitionDraftDsl;
 import io.sphere.sdk.products.attributes.EnumAttributeType;
 import io.sphere.sdk.products.attributes.LocalizedEnumAttributeType;
 import io.sphere.sdk.products.attributes.MoneyAttributeType;
+import io.sphere.sdk.products.attributes.ReferenceAttributeType;
 import io.sphere.sdk.products.attributes.SetAttributeType;
 import io.sphere.sdk.products.attributes.StringAttributeType;
 import io.sphere.sdk.producttypes.ProductType;
@@ -945,5 +949,54 @@ class ProductTypeSyncIT {
                 assertAttributesAreEqual(
                     productType.getAttributes(),
                     asList(withSetOfEnumsNew, withSetOfSetOfLEnumsNew)));
+  }
+
+  @Test
+  void sync_withProductTypeWithCategoryReference_ShouldAddNewAttributesToTheProductType() {
+    final ProductTypeSyncOptions productTypeSyncOptions =
+        ProductTypeSyncOptionsBuilder.of(CTP_TARGET_CLIENT).build();
+
+    AttributeDefinition referenceTypeAttr =
+        AttributeDefinitionBuilder.of(
+                "referenceTypeAttr",
+                LocalizedString.ofEnglish("referenceTypeAttr"),
+                SetAttributeType.of(ReferenceAttributeType.ofCategory()))
+            .isRequired(false)
+            .attributeConstraint(AttributeConstraint.NONE)
+            .isSearchable(false)
+            .build();
+    final ProductTypeDraft newProductTypeDraft =
+        ProductTypeDraft.ofAttributeDefinitionDrafts(
+            PRODUCT_TYPE_KEY_1,
+            PRODUCT_TYPE_NAME_1,
+            PRODUCT_TYPE_DESCRIPTION_1,
+            singletonList(AttributeDefinitionDraftBuilder.of(referenceTypeAttr).build()));
+
+    final ProductTypeSync productTypeSync = new ProductTypeSync(productTypeSyncOptions);
+    productTypeSync.sync(singletonList(newProductTypeDraft)).toCompletableFuture().join();
+
+    final ProductTypeDraft updatedProductTypeDraft =
+        ProductTypeDraft.ofAttributeDefinitionDrafts(
+            PRODUCT_TYPE_KEY_1,
+            PRODUCT_TYPE_NAME_1,
+            PRODUCT_TYPE_DESCRIPTION_1,
+            asList(
+                ATTRIBUTE_DEFINITION_DRAFT_1,
+                AttributeDefinitionDraftBuilder.of(referenceTypeAttr).build()));
+
+    productTypeSync.sync(singletonList(updatedProductTypeDraft)).toCompletableFuture().join();
+
+    final Optional<ProductType> updatedProductType =
+        getProductTypeByKey(CTP_TARGET_CLIENT, PRODUCT_TYPE_KEY_1);
+    assert updatedProductType.isPresent();
+
+    Optional<AttributeDefinition> newAttributeDefinition =
+        updatedProductType.get().getAttributes().stream()
+            .filter(
+                attributeDefinition ->
+                    attributeDefinition.getName().equals(ATTRIBUTE_DEFINITION_DRAFT_1.getName()))
+            .findAny();
+
+    assert newAttributeDefinition.isPresent();
   }
 }
